Move Compute Generic Parameter Specifics to compute type specifics pass
