+++ Il2CppInterop.Generator/Runners/DeobfuscationMapGenerator.cs
@@ -69,11 +69,6 @@ internal class DeobfuscationMapGeneratorRunner : IRunner
             Pass10CreateTypedefs.DoPass(rewriteContext);
         }
 
-        using (new TimingCookie("Computing generic parameter usage"))
-        {
-            Pass11ComputeGenericParameterSpecifics.DoPass(rewriteContext);
-        }
-
         using (new TimingCookie("Computing struct blittability"))
         {
             Pass12ComputeTypeSpecifics.DoPass(rewriteContext);
